Change store to fix bug where audio would restart at beginning
Create Playlist slice of state in the store.  The bug occured whenever I went to a Song Page or the Splash Page because those components called ajax requests that updated the song state in the store.  This, in turn, updated the AudioPlayer component that updated the SongBar. I think this rerendered the audio tag inside the SongBar, causing the music to reset.  I made sure that the SongBar only subscribes to the Playlist slice of state, and not the Songs slice of state directly.  When the Splash page tries to index its songs, or when a Song Page tries to get its song, the SongBar will not be affected.