Fix horrid bug where refreshing app messes up state
This partial jbuilder file for the user show view was used to create the window.currentUser that persists upon refreshes.  This kept some data of the current user, but not all: the user songIds were missing.  So when everything was refreshed, some pages that did not use the user.songIds worked fine, but when I went to the Users Page, the app crashed.  I fixed that by adding the new line that adds songIds.